+++ app/src/main/java/com/foobnix/ui2/fragment/PrefFragment2.java
@@ -136,9 +136,9 @@ public class PrefFragment2 extends UIFragment {
     public static final Pair<Integer, Integer> PAIR =
             new Pair<>(R.string.preferences, R.drawable.glyphicons_5_settings);
 
-    private static final String WWW_SITE = "https://librera.mobi";
-    private static final String WWW_BETA_SITE = "http://beta.librera.mobi";
-    private static final String WWW_WIKI_SITE = "https://librera.mobi/faq";
+    private static final String WWW_SITE = "https://github.com/380121850/LibreraReader";
+    private static final String WWW_BETA_SITE = "https://github.com/380121850/LibreraReader/releases";
+    private static final String WWW_WIKI_SITE = "https://github.com/380121850/LibreraReader";
     View section1, section2, section4, section8, section9, panelRecent, overlay,
             statusBarHack;
     TextView singIn, syncInfo, syncInfo2, syncHeader;